switched action to api route
